+++ src/object/isEmptyObj.ts
@@ -1,0 +1,4 @@
+export const isEmptyObj = (obj: object): boolean =>
+  obj &&
+  Object.keys(obj).length === 0 &&
+  Object.getPrototypeOf(obj) === Object.prototype;