+++ src/string/index.ts
@@ -3,3 +3,5 @@ export * from './stripcslashes';
 export * from './parseVariables';
 export * from './capitalizeFirstLetter';
 export * from './capitalizeEachWord';
+export * from './isAscii';
+export * from './shortenString';