+++ src/string/isAscii.ts
@@ -1,0 +1,1 @@
+export const isAscii = (text: string): boolean => /^[\x00-\x7F]*$/.test(text);